Fixed an issue with the query "Hi."
"Hi" would result to "H.I."--an acronym that we don't really need.
To handle this, we need to put everything in lowercase, then check if it finds a match (I find that this works most of the time).
If it doesn't, we can always rely on Dictionary::Fallback to handle it.
But just in case Dictionary::Fallback fails, we have DDG.isRelevant.