More relevant results in the dictionary plugin.
I removed Dictionary::Fallback and added useCanonical=true to Dictionary::Definition. This auto-corrects some of the spelling:

- brobdingnagian -> Brobdingnagian.
- cats -> cat.

However, it doesn't always do this. When we search "cacti," setting useCanonical=true will not correct it to "cactus." Why do we wan't to correct it? Because sometimes the plural form of something is not helpful. It would just say, "Plural form of cactus." So I thought that maybe we could add additional tags, not just to improve the way the plugin looks, but to also see which words are referenced in the definition. An example would be:

Plural form of <xref>cat</xref>.

Now we can find the xref tags and do something about it. For now, we convert it into a link e.g. "/?q=define+cat," but I think we can do better.