Improvements on the Dictionary plugin.
Some words such as "dictionaries" will just return "Plural form of <xref>dictionary</xref>," which isn't very helpful. But now the dictionary plugin will say that "dictionaries is the plural of dictionary" and the definitions below will be that of the word "dictionary." This is done by:

1. Checking if there is only one definition.
2. Checking if that definition matches a regexp i.e. /^(?:A )?plural (?:form )?of <xref>(\w+(?:\s\w+)*)<\/xref>/i

The pronunciation (both audio and text) will not exist for the plural form, so we get the one from the singular form e.g. "owls is the plural of owl (oul) (audio)."