InTheaters: Better image handling.
The images would change when we hover on the list of movies. Before, we only had one `img` tag, and we just replaced the `src` attribute whenever we hover on a movie. Now the problem with that is the browser does a request whenever we change the image. A solution to that is to create as many number of image tags as we have movies in the list, and just hide and show them accordingly i.e. if we have five movies in the list, we will create five `img` tags--four of them are hidden.

I added a tiny little feature, too. Images don't load unless you hover on them. This was inspired by Mika Tuupola's Lazy Load plugin.