Enabled caching on jQuery.
jQuery adds a "_={timestamp}" in the URL to prevent caching, but that doesn't work well on Duckpan.

e.g. It will try to request for "https://api.wordnik.com/v4/word.json/owl?_=1234567/definitions," and Wordnik thinks the word that we want is "owl?_=1234567." Setting $.ajaxSetup({cache: true}) fixes this.