Video: Do not trigger with "khan" or "khan academy".
We should skip queries which are in the form of "khan <trigger words>" and "khan academy <trigger words>" because we want the Khan Academy plugin to show us results instead. This change only checks if the remainder is "khan" or "khan academy," and that rejects queries such as "khan academy videos", but not "wrath of khan videos".