Quixey: Use HTTP instead of HTTPS for the image links.
I made a function called toHTTP which returns the HTTP version of icon_url. It gets the domain name of the image server that it finds in the platforms array, and it uses that to replace the domain name of the icon_url. It's not ideal, but it seems to work pretty well.