fix video.js parsing of .com searches
After an accidental search, I noticed that we're not properly parsing video
queries of the form 'youtube.com foo', only discarding 'youtube' and leaving
'.com' in the title. At first I thought it was an error in the trigger handler,
so I wrote a test (we needed one anyway) but it turns out it's just the js
that's got it wrong. Anyway, this fixes that :-)

see https://duckduckgo.com/?q=youtube.com+life+round+here