Forecast: rework triggering.
Intended effect is to require temperature trigger to be preceded by
"current" or followed by "in/at/for".. or both.

In pursuit of this, also simplified the way forecast skip words
are maintained. At least I think it is simpler.

Also eliminated `meteo` trigger as it seems pretty esoteric and
unhelpful.

Intended to address #1170.