Time: anchor and improve query regex.
This makes the querying much tighter by anchoring the regex to appear to
request at a particular place. It's still got some weirdness.
We're allowing for some seemingly unreasonable word orders, because
enumerating all of them would be ardous.. and likely wrong.

This does now require a connecting word (in, of, for, at) between time and
the place because otherwise it will likely trigger too wide.  This seems
to be in line with the previous version.