Expatistan: Display Result For Country Comparison
The Expatistan API will return a "abstract" result like "Cost of living in <b>London</b> is 23% higher than in <b>Paris</b>" for a country comparision.
Due to the regexp used to split the string based on the index (number) the firstline var containes no data when a country comparision is performed.
Simply check if we have an index (number) if not just return the firstline exactly how it is recieved from the API.