IndeedJobs.pm: Switch to deep trigger.
We're switching to deep triggers because it's difficult to extract the location from a query. We can understand location for some queries such as "javascript jobs in philadelphia" because of the "in" preposition, but we can do nothing for queries such as "javascript nyc jobs" because we don't have the functionality to get the location part of the query. For the mean time, we switch to a Deep trigger so that we can pass the right location to the Indeed API endpoint.

**What is a Deep trigger?**

A Deep trigger is triggering based on the SERP results instead of on the query. This means we can look in the search results, see indeed.com domains (such as indeed.com/q-Javascript-l-San-Francisco,-CA-jobs.html), extract the things that we want using regexp (in this case, "Javascript" and "San Francisco, CA"), and pass that along to the endpoint. Unfortunately, it isn't open source at the moment so it's a bit of a black box at the moment.