Fix #2323: Problems with Fandango
GoWatchIt's code has gotten a bit unwieldy and is due for another look, but the following fix works. So we have some code that strips some text from the `buy_line` or the `rent_line` without even checking if it should be stripped or not. This patch makes sure that such a thing doesn't happen again.