Fix #2316: Problem with loading animation
A keyframe with the same name has conflicted with Riffsy's keyframe. A better solution for this is to have gif detection integrated into the global image template (thinking of working on this soon!), but for now, we don't want it broken so here's a patch that renames the keyframe so that it doesn't conflict.