Fix handlebars MD parser to match coderwall's
This commit adds minor fixes to the Markdown Handlebars
helper. Specifically, the replace methods used to parse
italics and bold.

Coderwall's markdown parser seems to be based on Github's.
Although italics and bold don't result in any visual change
on their site, the parser is still marking up the <em> and
<strong> tags.

So, as of writing this, Coderwall will markup this syntax:

  *word*       -->  italic
  _word_       -->  italic
  __word__     -->  bold
  **word**     -->  bold
  `word`       -->  monospace/code
  ~~word~~     -->  strikethrough
  [word](url)  -->  link

They are also parsing ordered/unordered lists and headers.
My inclination is to not parse these, as it seems (anecdotally)
that this is not commonly used by users when they author their
descriptions, and, mainly, because rendering the lists and
headers will very quickly bloat div.cz-main and eat up a lot
of vertically space.

Coderwall does _NOT_ parsing images or tables.