Add custom handlebars block helper to substr HTML
{{#coderwall_substr_html}} is a custom block helper that trims
HTML text to a specific length. It does this by parsing the
context/string into HTML. The helper then iterates through each
text node, and then, if the textContent is longer than the
specified character limit, the text is trimmed to the closest
word. Once the limit is hit, all remaining nodes are exluded.

Ideally, this custom helper will be a temporary solution until the
built-in {{#condense}} block helper supports markup.